This makes unsynced type same as synced.