@gyuho 2016-01-05
Benchmarked before and after, and I don't see much difference:

```
BenchmarkWatchableStore:
[1]:
benchmark                                   old ns/op      new ns/op     delta
BenchmarkWatchableStoreUnsyncedCancel       626            662           +5.75%
BenchmarkWatchableStoreUnsyncedCancel-2     626            632           +0.96%
BenchmarkWatchableStoreUnsyncedCancel-4     627            656           +4.63%
BenchmarkWatchableStoreUnsyncedCancel-8     622            682           +9.65%
BenchmarkWatchableStoreSyncedCancel         1045902389     0.60          -100.00%
BenchmarkWatchableStoreSyncedCancel-2       0.30           0.59          +96.67%
BenchmarkWatchableStoreSyncedCancel-4       0.31           0.31          +0.00%
BenchmarkWatchableStoreSyncedCancel-8       0.64           0.34          -46.88%

benchmark                                   old allocs     new allocs     delta
BenchmarkWatchableStoreUnsyncedCancel       0              0              +0.00%
BenchmarkWatchableStoreUnsyncedCancel-2     0              0              +0.00%
BenchmarkWatchableStoreUnsyncedCancel-4     0              0              +0.00%
BenchmarkWatchableStoreUnsyncedCancel-8     0              0              +0.00%
BenchmarkWatchableStoreSyncedCancel         67             0              -100.00%
BenchmarkWatchableStoreSyncedCancel-2       0              0              +0.00%
BenchmarkWatchableStoreSyncedCancel-4       0              0              +0.00%
BenchmarkWatchableStoreSyncedCancel-8       0              0              +0.00%

benchmark                                   old bytes     new bytes     delta
BenchmarkWatchableStoreUnsyncedCancel       0             0             +0.00%
BenchmarkWatchableStoreUnsyncedCancel-2     0             0             +0.00%
BenchmarkWatchableStoreUnsyncedCancel-4     0             0             +0.00%
BenchmarkWatchableStoreUnsyncedCancel-8     0             0             +0.00%
BenchmarkWatchableStoreSyncedCancel         12088         0             -100.00%
BenchmarkWatchableStoreSyncedCancel-2       0             0             +0.00%
BenchmarkWatchableStoreSyncedCancel-4       0             0             +0.00%
BenchmarkWatchableStoreSyncedCancel-8       0             0             +0.00%


[2]:
benchmark                                   old ns/op     new ns/op     delta
BenchmarkWatchableStoreUnsyncedCancel       604           696           +15.23%
BenchmarkWatchableStoreUnsyncedCancel-2     608           651           +7.07%
BenchmarkWatchableStoreUnsyncedCancel-4     612           661           +8.01%
BenchmarkWatchableStoreUnsyncedCancel-8     760           701           -7.76%
BenchmarkWatchableStoreSyncedCancel         0.60          0.69          +15.00%
BenchmarkWatchableStoreSyncedCancel-2       0.30          0.29          -3.33%
BenchmarkWatchableStoreSyncedCancel-4       0.30          0.29          -3.33%
BenchmarkWatchableStoreSyncedCancel-8       0.60          0.62          +3.33%

benchmark                                   old allocs     new allocs     delta
BenchmarkWatchableStoreUnsyncedCancel       0              0              +0.00%
BenchmarkWatchableStoreUnsyncedCancel-2     0              0              +0.00%
BenchmarkWatchableStoreUnsyncedCancel-4     0              0              +0.00%
BenchmarkWatchableStoreUnsyncedCancel-8     0              0              +0.00%
BenchmarkWatchableStoreSyncedCancel         0              0              +0.00%
BenchmarkWatchableStoreSyncedCancel-2       0              0              +0.00%
BenchmarkWatchableStoreSyncedCancel-4       0              0              +0.00%
BenchmarkWatchableStoreSyncedCancel-8       0              0              +0.00%

benchmark                                   old bytes     new bytes     delta
BenchmarkWatchableStoreUnsyncedCancel       0             0             +0.00%
BenchmarkWatchableStoreUnsyncedCancel-2     0             0             +0.00%
BenchmarkWatchableStoreUnsyncedCancel-4     0             0             +0.00%
BenchmarkWatchableStoreUnsyncedCancel-8     0             0             +0.00%
BenchmarkWatchableStoreSyncedCancel         0             0             +0.00%
BenchmarkWatchableStoreSyncedCancel-2       0             0             +0.00%
BenchmarkWatchableStoreSyncedCancel-4       0             0             +0.00%
BenchmarkWatchableStoreSyncedCancel-8       0             0             +0.00%

```

/cc @xiang90 @heyitsanthony