@gyuho 2016-01-13
Closing this. Will be replaced by another PR to clean up `syncWatchers` method.